Refactor java file finder out of parser
